Modifying eatcpu.py to run with pbs python.
